Support CMake FetchContent
When pfs is used with the 'FetchContent' CMake functionality, there is
an issue with the usage of the 'CMAKE_PROJECT_NAME' parameter.

The parent project declares its own CMAKE_PROJECT_NAME, while the pfs
declares its own CMAKE_PROJECT_NAME. That is causing a collision, as pfs
does refer the CMAKE_PROJECT_NAME parameter inside its CMake files.

This patch removes referencing the CMAKE_PROJECT_NAME inside the CMake
files, making it possible to use the pfs project in the 'FetchContent'
context.